Cache negative results of equation attempts
