+++ kore/kore.cabal
@@ -275,6 +275,7 @@ library
     Kore.Debug
     Kore.Equation
     Kore.Equation.Application
+    Kore.Equation.DebugEquation
     Kore.Equation.Equation
     Kore.Equation.Registry
     Kore.Equation.Sentence
@@ -384,6 +385,7 @@ library
     Kore.Rewrite.Axiom.EvaluationStrategy
     Kore.Rewrite.Axiom.Identifier
     Kore.Rewrite.Axiom.Matcher
+    Kore.Rewrite.Axiom.MatcherData
     Kore.Rewrite.Axiom.Registry
     Kore.Rewrite.AxiomPattern
     Kore.Rewrite.ClaimPattern
