+++ kore/src/Kore/Attribute/Axiom/Constructor.hs
@@ -21,16 +21,10 @@ import Prelude.Kore
 -- | @Constructor@ represents the @constructor@ attribute for axioms.
 newtype Constructor = Constructor {isConstructor :: Bool}
     deriving stock (Eq, GHC.Generic, Ord, Show)
-
-instance SOP.Generic Constructor
-
-instance SOP.HasDatatypeInfo Constructor
-
-instance Debug Constructor
-
-instance Diff Constructor
-
-instance NFData Constructor
+    deriving anyclass (SOP.Generic, SOP.HasDatatypeInfo)
+    deriving anyclass (Debug, Diff)
+    deriving anyclass (NFData)
+    deriving anyclass (Hashable)
 
 instance Default Constructor where
     def = Constructor False
