+++ kore/src/Kore/Log/Registry.hs
@@ -34,7 +34,7 @@ import Data.Text (
     Text,
  )
 import qualified Data.Text as Text
-import Kore.Equation.Application (
+import Kore.Equation.DebugEquation (
     DebugApplyEquation,
     DebugAttemptEquation,
  )
